+++ uplot/__init__.py
@@ -1,4 +1,4 @@
-from uplot.info import __version__, __author__, __email__
+from uplot.info import __version__, __author__, __email__  # noqa: F401
 
 # engine managing
 import uplot.engine as engine
@@ -13,7 +13,37 @@ from uplot.plot import figure
 import uplot.color as color
 
 # common types
-from uplot.utype import LineStyle, MarkerStyle, AspectMode, Colormap
+from uplot.utype import LineStyle, MarkerStyle, AspectMode, AxisScale, Colormap
 
 # settings
 from uplot.default import DEFAULT
+
+
+__all__ = [
+    
+    # modules
+
+    'engine',
+    'color',
+
+    # interface
+
+    'IFigure', 
+    'IPlotEngine',
+    
+    # functions
+    
+    'figure',
+
+    # types
+   
+    'LineStyle', 
+    'MarkerStyle', 
+    'AspectMode',
+    'AxisScale', 
+    'Colormap',
+
+    # variables / constants
+
+    'DEFAULT'
+]