+++ uplot/engine/MatplotFigure.py
@@ -9,7 +9,7 @@ import uplot.color as ucolor
 import uplot.utool as utool
 import uplot.plugin as plugin
 
-from uplot.interface import IFigure, LineStyle, MarkerStyle, AspectMode, Colormap
+from uplot.interface import IFigure, LineStyle, MarkerStyle, AspectMode, AxisScale, Colormap
 from uplot.engine.MatplotEngine import MatplotEngine
 from uplot.utool import Interpolator
 from uplot.default import DEFAULT
@@ -30,14 +30,12 @@ class MatplotFigure(IFigure):
         return self._is_3d
 
     def __init__(self, engine: MatplotEngine, width: int, aspect_ratio: float):
-        self._engine = engine
-        self._is_3d = None
-        self._axis: engine.plt.Axes = None
+        from matplotlib.figure import Figure
 
         # temporary styling (no global effect):
         # https://matplotlib.org/stable/users/explain/customizing.html
         with engine.plt.style.context(DEFAULT.style):
-            self._fig: engine.plt.Figure = engine.plt.figure(dpi=engine.SHOWING_DPI, layout='constrained')
+            self._fig: Figure | None = engine.plt.figure(dpi=engine.SHOWING_DPI, layout='constrained')
             # constrained layout automatically adjusts subplots so that decorations like tick labels,
             # legends, and colorbars do not overlap, while still preserving the logical layout requested by the user.
             # constrained layout is similar to Tight layout, but is substantially more flexible.
@@ -45,7 +43,9 @@ class MatplotFigure(IFigure):
             self._fig.set_figwidth(width / engine.SHOWING_DPI)
             self._fig.set_figheight(aspect_ratio*(width / engine.SHOWING_DPI))
 
+        self._engine = engine
         self._color_scroller = ucolor.ColorScroller()
+        self._is_3d = None
         self._init_axis(is_3d=False)
         self._bars = [ ]
 
@@ -56,7 +56,7 @@ class MatplotFigure(IFigure):
                    color       : str | None = None,
                    line_style  : LineStyle | None = None,
                    marker_style: MarkerStyle | None = None,
-                   marker_size : int | None = None,
+                   marker_size : float | None = None,
                    opacity     : float = 1.0,
                    legend_group: str | None = None,
                    **kwargs) -> IFigure:
@@ -99,7 +99,7 @@ class MatplotFigure(IFigure):
                       name        : str | None = None,
                       color       : str | list[str] | None = None,
                       marker_style: MarkerStyle | None = None,
-                      marker_size : int | None = None,
+                      marker_size : float | None = None,
                       opacity     : float = 1.0,
                       legend_group: str | None = None,
                       **kwargs) -> IFigure:
@@ -199,6 +199,8 @@ class MatplotFigure(IFigure):
                         interpolation_range: int = 100,
                         legend_group : str | None = None,
                         **kwargs) -> IFigure:
+        assert self._fig is not None, 'figure is closed'
+
         # check if x is a custom object and a plugin is available
         if plugin.plot(plot_method=self.surface3d,
                        x=x, y=y, z=z,
@@ -229,6 +231,9 @@ class MatplotFigure(IFigure):
                                           interpolation_range=interpolation_range)
 
         axis = self._init_axis(is_3d=True)
+
+        from mpl_toolkits.mplot3d import Axes3D
+        assert isinstance(axis, Axes3D)
 
         cmap = self.engine.mpl.colormaps[colormap.lower()]
 
@@ -261,8 +266,8 @@ class MatplotFigure(IFigure):
             y = x
             x = np.arange(len(y))
         else:
-            assert len(x) == len(y), 'the length of the input arrays must be the same'
             y = np.asarray(y)
+            assert len(x) == len(y), 'the length of the input arrays must be the same'
 
         # init color
         if color is None:
@@ -338,6 +343,8 @@ class MatplotFigure(IFigure):
     def legend(self, show: bool = True,
                      equal_marker_size: bool = True,
                      **kwargs) -> IFigure:
+        assert self._fig is not None, 'figure is closed'
+
         if not self._axis:
             return self
 
@@ -349,7 +356,7 @@ class MatplotFigure(IFigure):
         if equal_marker_size:
             from matplotlib.legend_handler import HandlerPathCollection, HandlerLine2D
             from matplotlib.collections import PathCollection
-            from matplotlib.pyplot import Line2D
+            from matplotlib.lines import Line2D
 
             def updatescatter(handle, orig):
                 handle.update_from(orig)
@@ -381,7 +388,7 @@ class MatplotFigure(IFigure):
         return self
 
     def grid(self, show: bool = True) -> IFigure:
-        self._axis.grid(visible=show)
+        self._axis.grid(visible=show, which='both')
         return self
 
     def xlabel(self, text: str) -> IFigure:
@@ -394,6 +401,8 @@ class MatplotFigure(IFigure):
 
     def zlabel(self, text: str) -> IFigure:
         if self.is_3d:
+            from mpl_toolkits.mplot3d import Axes3D
+            assert isinstance(self._axis, Axes3D)
             self._axis.set_zlabel(zlabel=text)
         return self
 
@@ -410,7 +419,23 @@ class MatplotFigure(IFigure):
     def zlim(self, min_value: float | None = None,
                    max_value: float | None = None) -> IFigure:
         if self.is_3d:
+            from mpl_toolkits.mplot3d import Axes3D
+            assert isinstance(self._axis, Axes3D)
             self._axis.set_zlim(bottom=min_value, top=max_value)
+        return self
+
+    def xscale(self, scale: AxisScale, base: float = 10) -> IFigure:
+        if scale == 'linear':
+            self._axis.set_xscale(scale)
+        elif scale == 'log':
+            self._axis.set_xscale(scale, base=base)
+        return self
+
+    def yscale(self, scale: AxisScale, base: float = 10) -> IFigure:
+        if scale == 'linear':
+            self._axis.set_yscale(scale)
+        elif scale == 'log':
+            self._axis.set_yscale(scale, base=base)
         return self
 
     def current_color(self) -> str:
@@ -429,17 +454,24 @@ class MatplotFigure(IFigure):
         return self
 
     def as_image(self) -> ndarray:
+        assert self._fig is not None, 'figure is closed'
+
         fig = self._fig
 
         fig.set_dpi(self.engine.SAVING_DPI)
 
         fig.canvas.draw()
-        image = np.frombuffer(fig.canvas.tostring_rgb(), dtype=np.uint8)
+
+        from matplotlib.backends.backend_agg import FigureCanvasAgg
+        assert isinstance(fig.canvas, FigureCanvasAgg)
+
+        image = np.frombuffer(fig.canvas.buffer_rgba(), dtype=np.uint8)
 
         w, h = fig.canvas.get_width_height()
-        return image.reshape([h, w, 3])
+        return image.reshape([h, w, 4])
 
     def save(self, filename: str):
+        assert self._fig is not None, 'figure is closed'
         self._fig.savefig(filename, dpi=self.engine.SAVING_DPI)
 
     def close(self):
@@ -447,6 +479,8 @@ class MatplotFigure(IFigure):
         self._fig = None
 
     def show(self, block: bool=True):
+        assert self._fig is not None, 'figure is closed'
+
         if self.engine.is_ipython_backend:
             # there are two ways for consistent figure visualization in jupyter
             #    1. call `%matplotlib ...` at the notebook start.
@@ -468,6 +502,8 @@ class MatplotFigure(IFigure):
 
 
     def _init_axis(self, is_3d: bool):
+        assert self._fig is not None, 'figure is closed'
+
         if self.is_3d == is_3d:
             # axis already initialized
             return self._axis
@@ -481,7 +517,7 @@ class MatplotFigure(IFigure):
             self._axis = self._fig.add_subplot(projection=projection)
 
         self._is_3d = is_3d
-        self._axis.grid(visible=True) # show grid by default
+        self._axis.grid(visible=True, which='both') # show grid by default
 
         if is_3d:
             # sync axis and figure color