+++ uplot/engine/plotly/axis_range.py
@@ -1,6 +1,8 @@
 import numpy as np
 from typing import Literal
 
+from uplot.engine.plotly.scale import get_scale
+
 
 def estimate_axis_range(figure,
                         axis: Literal['x', 'y', 'z'],
@@ -32,13 +34,15 @@ def estimate_axis_range(figure,
     minmax = minmax_estimate(data_minmax)
 
     # estimate min/max from range
-    axis = {
+    axis_name: str = {
         'x': 'xaxis',
         'y': 'yaxis'
     }[axis]
 
-    axis_range = figure.layout[axis]['range']
+    axis_range = figure.layout[axis_name]['range']
     if axis_range is not None:
+        if get_scale(figure, axis) == 'log':
+            axis_range = 10**np.asarray(axis_range)
         minmax = minmax_estimate([ minmax, *axis_range ])
 
     return minmax