+++ uplot/utool/__init__.py
@@ -1,4 +1,18 @@
 from uplot.utool.param import unpack_param
 from uplot.utool.image import image_range, image_encode_base64
-from uplot.utool.StrEnum import StrEnum
 from uplot.utool.grid import array_to_grid, Interpolator
+
+
+__all__ = [ 
+
+    # functions
+
+    'unpack_param',
+    'image_range',
+    'image_encode_base64',
+    'array_to_grid',
+
+    # types
+
+    'Interpolator'
+]