+++ uplot/utype.py
@@ -1,10 +1,16 @@
 from typing import Literal
-from uplot.utool import StrEnum
 
 
-class AspectMode(StrEnum):
-    AUTO = 'auto'
-    EQUAL = 'equal' # equal scale for all axis
+AxisScale = Literal[
+    'linear',
+    'log',
+]
+
+
+AspectMode = Literal[
+    'auto',
+    'equal' # equal scale for all axis
+]
 
 
 LineStyle = Literal[